.github: [Enh] Update & Clean windows build workflow to be faster (#21)
We remove the CMake and Vcpkg install since the provided versionsfrom
windows-latest images is "sufficient". We keep the cache but use a
manual one instead of using the run vcpkg action to do it.

When LLVM will be updated to >16, we can remove the install and setup
LLVM steps to make it event faster.