.github: [Fix] Define __cpp_consteval macro on clang
Problem:
When using clang on windows, the MSVC STL is always used. When trying to
use `std::source_location`, the compiler crashes and a STL warning is
raised saying `The contents of <source_location> are available only with
C++20 consteval support`.

Solution:
The content of source_location in 14.36 (installed on github runners) is
only available with the __cpp_consteval macro defined, so define it in
the comile flags since it's not done by clang.