spark/patterns: [Fix] Slot not disconnecting from signal
Problem:
When calling the disconnect method on a slot, the isConnected() method
still returns true.

Cause:
The Signal::disconenct method performs a search in the conenction map by
key but it's a map of slots. It is always unable to find any slot,
leading to never disconnect anything.

Solution:
Perform the search with a srd::ranges::find_if, and specify we want to
search by key.