spark/rtti, vcpkg.json: [Fix] Boost preprocessor headers not linked to target
Problem:
When using rtti in a target thats not using vcpkg library, it does not
compiles due to the boostpp headers not being found.

Solution:
Link boost header in the CMake. To allow that, we need to make
FindBoost.cmake work. It requires the boost/config.hpp to find for the
library.