spark/engine: [Fix] Crash due to assertion failed on close
Problem:
When closing a game, the application crashes because onDestroyed() is
called 2 times on the GameObject.

Solution:
Fixing this bug entirely will not be easy sonce it will require to
delete and recreate swene hieracy every time it's loaded or unloaded.
So, for now, we replace the asserts in the onSpawn() and onDestroyed()
methods by early returns.