spark/render/vk: [Fix] Created structured buffer data is nullptr
Problem:
When updating a descriptor set with a structured buffer, the validation
layers warns about pBufferData being nullptr.

Cause:
The pointer to the data is assigned before assigning values to the
vector, which becomes the wrong value when the vector reallocates.

Solution:
Assign the pointer after the vector is completly assigned.