spark/rtti: [Fix] Database::Get() always returns nullptr
Problem:
When calling RttiDatabase::Get() with a registered class name, it
returns nullptr.

Cause:
The name was not correctly registered in the Rtti object, value of the
map and was an unknown string.

Solution:
Correctly register the name when creating Rtti objects, passing it as
ctor arg instead of using the tpl struct.