several: [Enh] Merge core and engine packages
Previously, we had a cyclic dependency between core, which needed scenes
and engine, which needed core for rendering. It was working as is
because core dependency in engine was header only and the transitively
includes did the job.

Now, we need to add serialization, which is everywhere and it breaks
everything. So, we merge both packages since it was already not
independent at all.