spark/core: [Fix] GameObjects not initialized when created at runtime
Problem:
When constructing a new gameobject at runtime (after scene is loaded),
the abstract onSpawn() method is never called, leading to m_initialierd
being false and onDestroyed never being called.

Cause:
The onSpawn() method was called on the GameObject, not the
AbstractGameObject one, which is hidden.

Solution:
Do the cast! This should be really tested someday !